Refactored Home UI folder to align with BaseHome and BaseHomeViewModel
