+++ app/src/main/java/com/fwrdgrp/wordapp/ui/home/CompleteWordFragment.kt
@@ -4,5 +4,4 @@ import androidx.fragment.app.viewModels
 
 class CompleteWordFragment : BaseHomeManageFragment() {
     override val viewModel: CompleteWordViewModel by viewModels()
-
 }
